refactor: split logic in separate files
